requirements.txt
psutil
pyjwt
protobuf-decoder
requests
urllib3
google
pycryptodome
protobuf
python-cfonts
telebot